[fact] Put all class representatives method into a single class #29
